Make PrimitiveTypeConverter static
It was only used as a pseudo-Singleton in private vars or in a static variable anyways and only uses static data for the decision process.
So to avoid any instance null checks and such I first wanted to turn it into a proper Singleton, but then realized it can be fully static as well, which is easier.