Add Deserialization FlatArrayHandler
FlatArrayHandler works on Lists that have the FlatArray Attribute.
The handler is effectively just a list handler, except, that it does not work on the Children of the node given to it, but directly on the collection of nodes given to it.

The Unit Test "NewDeserializerTest" Tests both a regular Flat List as well as a Flat List of Tuples.