Fix and Finish new Deserializers (#25)
* Make PrimitiveTypeConverter static

It was only used as a pseudo-Singleton in private vars or in a static variable anyways and only uses static data for the decision process.
So to avoid any instance null checks and such I first wanted to turn it into a proper Singleton, but then realized it can be fully static as well, which is easier.

* Introduce Property for "size"-array entryname in FileDBSerializerOptions

Thus when it gets reused (eg. in the deserializer) there are no typos.

* Make SelectSingleNode nullable

Since the function is specified to return null when no matching node is found, it should be defined as nullable.

* Update ReferenceArray Deserialization Handler to handle array size

The ReferenceArrayHandler now reads the size attrib from the array node and checks if it matches the actual amount of children in the array.

* Add Deserialization ListHandler

Add the Deserialization ListHandler, which creates a Generic List<TargetType> via Reflection. Luckily since the item handler just returns object? anyways, we don't have to cast the created list object to its generic type, but just to IList, and can add the objects to that.

Expanded the Unit Test "NewDeserializerTest" by both Reference and Primitive Lists (added these elements to the DeserObject class).

* Add Deserialization TupleHandler

Added TupleHandler, which also handles nested Tuples (that are separately wrapped in a Tag).
Made List of Tuples correctly group Tuple items as they are flattened in the List.
Made ReferenceArrayHandler correctly unpack the extra tag around Tuple children.

Unit Test "NewDeserializerTest" handles following cases:
- regular Tuple
- Nested Tuple
- List of Tuples
- Array of Tuples

* Add Deserialization FlatArrayHandler

FlatArrayHandler works on Lists that have the FlatArray Attribute.
The handler is effectively just a list handler, except, that it does not work on the Children of the node given to it, but directly on the collection of nodes given to it.

The Unit Test "NewDeserializerTest" Tests both a regular Flat List as well as a Flat List of Tuples.