Empty ReferenceArrays export without size tag
Empty Reference arrays should export without size tag (eg. AnimalHandler->HerdAreas->None->DangerPoints). The deserializer already accepts empty/self-closing tags as Reference Arrays, but currently empty Reference Arrays export with a <size> tag.

This used to be intended behaviour to allow Reference Arrays with a <size>0</size> tag, and the empty array would be done with a null ReferenceArray and a corresponding FileDBSerializerOption, but that leads to the effect that there is a discrepancy when deserializing data and directly serializing it again.

This change makes it so that a ReferenceArray tag with <size>0</size> can never be exported, and reading a ReferenceArray with only 1 child in the tag throws an error.

A UnitTest to confirm that an error is thrown when deserializing a ReferenceArray with exactly 1 child tag was added (ErrorOnInvalidReferenceArray).
The SkipSimpleNullValues UnitTest was adapted to match the new default behaviour.

An additional UnitTest was added to ensure that deserializing and then serializing an empty ReferenceArray keeps the xml document the same.

Also, a mistake with the FileDBDocumentDeserializer was resolved, as its top layer handler should act like the ReferenceTypeHandler does (iterate over the FileDBNodes), but it acted differently (iterated over target type properties).