Make sure empty attribs are recognized as such
When turning an string or file to an XmlDocument, empty, non-selfclosing tags may have no children, but are an XmlElement that has the property IsEmpty set to false (would be true on self-closing elements). This scenario was previously not respected, making empty attribs be read as Tags and then causing deserialization to fail.

Added unit test "DeSerializeEmptyPrimitiveArray" to verify result.